requirements.txt
flask
gunicorn
google-genai
Pillow